JUS-41 refactoring code
Refactored the code a bit and removed excess code from the branch
