--- server/api/test.route.js
@@ -1,7 +1,0 @@
-import express from "express";
-
-const router = express.Router();
-
-router.route("/").get((req, res) => res.send("Hello World!"));
-
-export {router};
