+++ server/auth/serialize.js
@@ -1,5 +1,5 @@
-import { Customer } from "../api/models/customer.model.js";
-import { EventOrganizer } from "../api/models/eventOrganizer.model.js";
+import { Customer } from "../models/customer.model.js"
+import { EventOrganizer } from "../models/eventOrganizer.model.js"
 
 function configSerialization(passport){ 
     passport.serializeUser((user, done) => done(null, user.id))
