+++ server/database/DBReset.js
@@ -1,7 +1,6 @@
 import mongoose from 'mongoose'
-import { Customer } from '../models/customer.model.js'; 
-import { EventOrganizer } from '../models/eventOrganizer.model.js';
-import { Event } from '../models/event.model.js';
+import { Customer } from '../models/customer.model.js'
+import { EventOrganizer } from '../models/eventOrganizer.model.js'
 import dotenv from 'dotenv';
 dotenv.config();
