+++ server/database/DBconnect.js
@@ -1,5 +1,5 @@
 import mongoose from 'mongoose'
-import dotenv from 'dotenv';
+import dotenv from 'dotenv'
 dotenv.config();
