+++ server/models/schemas/bidInfo.schema.js
