+++ server/routes/customer/login.js
@@ -1,5 +1,5 @@
-import express from 'express';
-import mongoose from 'mongoose';
+import express from 'express'
+import mongoose from 'mongoose'
 
 const router = express.Router();
